requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
stripe
pytest
httpx
aiosqlite
email-validator
greenlet
bcrypt==4.0.1
